+++ src/inu/hardware/mmwave/mr24hpc1.py
@@ -2,7 +2,6 @@
 Seeed Studio MR24HPC1 24GHz Radar Sensor
 
 User manual: https://files.seeedstudio.com/wiki/mmWave-radar/MR24HPC1_User_Manual-V1.5.pdf
-Quick start: https://files.seeedstudio.com/wiki/mmWave-radar/MR24HPC1_Quick_Setup_Template-V1.0.pdf
 """
 import asyncio
 import logging
@@ -44,23 +43,35 @@ class ControlCodes:
     NO_PERSON_INQ = b'\x80\x8A'
     PROXIMITY_INQ = b'\x80\x8B'
 
+    # Custom mode
+    CUSTOM_MODE = b'\x05\x09'
+    END_CUSTOM_MODE = b'\x05\x0A'
+    CUSTOM_MODE_QUERY = b'\x05\x89'
 
-class Scene:
-    LIVING_ROOM = b'\x01'
-    BEDROOM = b'\x02'
-    BATHROOM = b'\x03'
-    AREA = b'\x04'
+    # Custom mode - set
+    EXISTENCE_THRESHOLD = b'\x08\x08'
+    MOTION_THRESHOLD = b'\x08\x09'
+    PERCEPTION_BOUNDARY = b'\x08\x0A'
+    MOTION_BOUNDARY = b'\x08\x0B'
+    MOTION_TRG_TIME = b'\x08\x0C'
+    IDLE_TRG_TIME = b'\x08\x0D'
+    NO_PERSON_TRG_TIME = b'\x08\x0E'
 
-    @staticmethod
-    def from_room_size(room_size: int):
-        if room_size >= 5:
-            return Scene.LIVING_ROOM
-        elif room_size == 4:
-            return Scene.BEDROOM
-        elif room_size == 3:
-            return Scene.AREA
-        else:
-            return Scene.BATHROOM
+    # Custom mode - query
+    EXISTENCE_THRESHOLD_INQ = b'\x08\x88'
+    MOTION_THRESHOLD_INQ = b'\x08\x89'
+    PERCEPTION_BOUNDARY_INQ = b'\x08\x8A'
+    MOTION_BOUNDARY_INQ = b'\x08\x8B'
+    MOTION_TRG_TIME_INQ = b'\x08\x8C'
+    IDLE_TRG_TIME_INQ = b'\x08\x8D'
+    NO_PERSON_TRG_TIME_INQ = b'\x08\x8E'
+
+
+class CustomMode:
+    MODE_1 = b'\x01'
+    MODE_2 = b'\x02'
+    MODE_3 = b'\x03'
+    MODE_4 = b'\x04'
 
 
 class Sensitivity:
@@ -78,26 +89,50 @@ class Sensitivity:
             return Sensitivity.LOW
 
 
+class Boundaries:
+    # 0.5m to 5.0m
+    M_05 = b'\x01'
+    M_10 = b'\x02'
+    M_15 = b'\x03'
+    M_20 = b'\x04'
+    M_25 = b'\x05'
+    M_30 = b'\x06'
+    M_35 = b'\x07'
+    M_40 = b'\x08'
+    M_45 = b'\x09'
+    M_50 = b'\x0A'
+
+    @staticmethod
+    def from_value(value: float):
+        if value < 0.75:
+            return Boundaries.M_05
+        elif value < 1.25:
+            return Boundaries.M_10
+        elif value < 1.75:
+            return Boundaries.M_15
+        elif value < 2.25:
+            return Boundaries.M_20
+        elif value < 2.75:
+            return Boundaries.M_25
+        elif value < 3.25:
+            return Boundaries.M_30
+        elif value < 3.75:
+            return Boundaries.M_35
+        elif value < 4.25:
+            return Boundaries.M_40
+        elif value < 4.75:
+            return Boundaries.M_45
+        else:
+            return Boundaries.M_50
+
+
 class Data:
+    FRAME_HEADER = b'\x53\x59'
+    FRAME_FOOTER = b'\x54\x43'
     NO_DATA = b'\x0F'
 
 
-class NoBodyDelay:
-    NONE = b'\x00'
-    SEC_10 = b'\x01'
-    SEC_30 = b'\x02'
-    MIN_1 = b'\x03'
-    MIN_2 = b'\x04'
-    MIN_5 = b'\x05'
-    MIN_10 = b'\x06'
-    MIN_30 = b'\x07'
-    HR_1 = b'\x08'
-
-
 class Mr24hpc1(RadarSensor):
-    FRAME_HEADER = b'\x53\x59'
-    FRAME_FOOTER = b'\x54\x43'
-
     def __init__(self, uart_index: int = 0):
         super().__init__()
         self.logger = logging.getLogger('inu.hw.mr24hpc1')
@@ -118,13 +153,13 @@ class Mr24hpc1(RadarSensor):
     def send_cmd(self, ctrl_code: bytes, data: bytes):
         self.logger.info(f"Sending command: {self.format_hex(ctrl_code)} :: {self.format_hex(data)}")
 
-        payload = bytearray(self.FRAME_HEADER)
+        payload = bytearray(Data.FRAME_HEADER)
         payload.extend(ctrl_code)
         payload.extend(len(data).to_bytes(2, 'big'))
         payload.extend(data)
         checksum = sum(payload) & 0xFF
         payload.extend(checksum.to_bytes(1, 'big'))
-        payload.extend(self.FRAME_FOOTER)
+        payload.extend(Data.FRAME_FOOTER)
 
         self.uart.write(payload)
         self.uart.flush()
@@ -135,7 +170,7 @@ class Mr24hpc1(RadarSensor):
         if header is None:
             return
 
-        if header != self.FRAME_HEADER:
+        if header != Data.FRAME_HEADER:
             self.logger.error(f"Invalid frame header ({self.format_hex(header)})")
             self.flush_input()
             return
@@ -161,7 +196,7 @@ class Mr24hpc1(RadarSensor):
         frame.extend(footer)
 
         # Validate footer
-        if footer != self.FRAME_FOOTER:
+        if footer != Data.FRAME_FOOTER:
             self.logger.error(f"Invalid frame footer ({self.format_hex(frame)})")
             return
 
@@ -243,13 +278,10 @@ class Mr24hpc1(RadarSensor):
 
             if data == b'\x00':
                 motion = None
-                self.logger.info(f"Motion: none")
             elif data == b'\x01':
                 motion = False
-                self.logger.info(f"Motion: idle")
             elif data == b'\x02':
                 motion = True
-                self.logger.info(f"Motion: active")
             else:
                 self.logger.error(f"Unknown motion state: {data}")
                 return
@@ -305,6 +337,132 @@ class Mr24hpc1(RadarSensor):
             if self.radar.direction != direction:
                 self.radar.direction = direction
 
+        elif ctrl_code == ControlCodes.CUSTOM_MODE_QUERY:
+            if data_len != 1:
+                self.logger.error(f"Invalid data length for custom mode query: {data_len}")
+                return
+
+            self.logger.info(f"Custom mode: {data[0]}")
+
+        elif ctrl_code == ControlCodes.CUSTOM_MODE:
+            if data_len != 1:
+                self.logger.error(f"Invalid data length for custom mode: {data_len}")
+                return
+
+            self.logger.info(f"Entered custom mode {data[0]}")
+
+        elif ctrl_code == ControlCodes.END_CUSTOM_MODE:
+            if data_len != 1:
+                self.logger.error(f"Invalid data length for custom mode end: {data_len}")
+                return
+
+            self.logger.info(f"Custom mode settings written")
+
+        elif ctrl_code == ControlCodes.EXISTENCE_THRESHOLD or ctrl_code == ControlCodes.EXISTENCE_THRESHOLD_INQ:
+            if data_len != 1:
+                self.logger.error(f"Invalid data length for existence threshold: {data_len}")
+                return
+
+            self.logger.info(f"Existence threshold: {data[0]}")
+
+        elif ctrl_code == ControlCodes.MOTION_THRESHOLD or ctrl_code == ControlCodes.MOTION_THRESHOLD_INQ:
+            if data_len != 1:
+                self.logger.error(f"Invalid data length for motion threshold: {data_len}")
+                return
+
+            self.logger.info(f"Motion trigger threshold: {data[0]}")
+
+        elif ctrl_code == ControlCodes.PERCEPTION_BOUNDARY or ctrl_code == ControlCodes.PERCEPTION_BOUNDARY_INQ:
+            if data_len != 1:
+                self.logger.error(f"Invalid data length for perception boundary: {data_len}")
+                return
+
+            self.logger.info(f"Perception boundary: {data}")
+
+        elif ctrl_code == ControlCodes.MOTION_BOUNDARY or ctrl_code == ControlCodes.MOTION_BOUNDARY_INQ:
+            if data_len != 1:
+                self.logger.error(f"Invalid data length for motion boundary: {data_len}")
+                return
+
+            self.logger.info(f"Motion boundary: {data}")
+
+        elif ctrl_code == ControlCodes.MOTION_TRG_TIME or ctrl_code == ControlCodes.MOTION_TRG_TIME_INQ:
+            if data_len != 4:
+                self.logger.error(f"Invalid data length for motion trigger time: {data_len}")
+                return
+
+            time_ms = int.from_bytes(data, 'big')
+            self.logger.info(f"Motion trigger time: {time_ms} ms")
+
+        elif ctrl_code == ControlCodes.IDLE_TRG_TIME or ctrl_code == ControlCodes.IDLE_TRG_TIME_INQ:
+            if data_len != 4:
+                self.logger.error(f"Invalid data length for idle trigger time: {data_len}")
+                return
+
+            time_ms = int.from_bytes(data, 'big')
+            self.logger.info(f"Idle trigger time: {time_ms} ms")
+
+        elif ctrl_code == ControlCodes.NO_PERSON_TRG_TIME or ctrl_code == ControlCodes.NO_PERSON_TRG_TIME_INQ:
+            if data_len != 4:
+                self.logger.error(f"Invalid data length for no-person trigger time: {data_len}")
+                return
+
+            time_ms = int.from_bytes(data, 'big')
+            self.logger.info(f"No-person trigger time: {time_ms} ms")
+
         # Anything else - not a big deal if we don't handle other control codes
         else:
             self.logger.warning(f"Unhandled control code: {self.format_hex(ctrl_code)}")
+
+    async def calibrate(self, cfg: dict):
+        """
+        "std": {
+                "existence_threshold": 33,
+                "motion_threshold": 4,
+                "perception_boundary": 2.5,
+                "motion_boundary": 2.5,
+                "time_for_motion": 150,
+                "time_for_idle": 2500,
+                "time_for_absence": 5000
+            },
+        """
+
+        def get_cfg(key: str, default=None):
+            if key not in cfg:
+                return default
+            return cfg[key]
+
+        # We always write to and use custom mode 1. If the template changes, we'll re-write the values for MODE_1.
+        self.send_cmd(ControlCodes.CUSTOM_MODE, CustomMode.MODE_1)
+        self.send_cmd(
+            ControlCodes.EXISTENCE_THRESHOLD,
+            get_cfg("existence_threshold", 33).to_bytes(1, 'big')
+        )
+        self.send_cmd(
+            ControlCodes.MOTION_THRESHOLD,
+            get_cfg("motion_threshold", 4).to_bytes(1, 'big')
+        )
+        self.send_cmd(
+            ControlCodes.PERCEPTION_BOUNDARY,
+            Boundaries.from_value(get_cfg("perception_boundary", 2.5))
+        )
+        self.send_cmd(
+            ControlCodes.MOTION_BOUNDARY,
+            Boundaries.from_value(get_cfg("motion_boundary", 2.5))
+        )
+        self.send_cmd(
+            ControlCodes.MOTION_TRG_TIME,
+            get_cfg("time_for_motion", 150).to_bytes(4, 'big')
+        )
+        self.send_cmd(
+            ControlCodes.IDLE_TRG_TIME,
+            get_cfg("time_for_idle", 3000).to_bytes(4, 'big')
+        )
+        self.send_cmd(
+            ControlCodes.NO_PERSON_TRG_TIME,
+            get_cfg("time_for_absence", 5000).to_bytes(4, 'big')
+        )
+        self.send_cmd(ControlCodes.END_CUSTOM_MODE, Data.NO_DATA)
+
+        # Should return 0x01
+        self.send_cmd(ControlCodes.CUSTOM_MODE_QUERY, Data.NO_DATA)